Made transfer of files > 4GB possible.

+++ dirsenderthread.cpp
@@ -80,7 +80,7 @@ void DirSenderThread::run()
     QByteArray data;
     QDataStream stream(&data, QIODevice::WriteOnly);
     stream.setVersion(QDataStream::Qt_4_5);
-    stream << (quint32)document.toByteArray().size();
+    stream << (quint64)document.toByteArray().size();
     data.append(document.toByteArray());
 
     socket.write(data);
